Add filtering and exclusion features to the schedule
This commit introduces several classes for filtering lectures based on various criteria such as subject, classroom size, etc. These classes implement the Filter and Exclusion interface to provide test methods which filter lectures according to specific requirements. Moreover, a static method was added to the Filter interface to create a new Filter object based on the given type and requirement.
Also, methods related to exclusions are added in the Schedule class. Multiple changes were made in the Schedule class to cater for these new features such as inclusion of lectures and checking if a lecture is excluded.
In addition, modifications were made to the Classroom and ValidityPeriod classes to collaborate with the filtering and exclusion strategies.
The motivation behind these changes is to support dynamic and versatile filtering of lectures, which enhances the usability of the schedule application.