Refactor filter creation in Filter class
Changed the filter creation method in the `Filter` class to now use an enum `FilterType` instead of a String to specify the type of filter. This eliminates the possibilities of typos and encourages usage of pre-defined filter types. Added `FilterType` enum to define types of filters. This code enhancement boosts type safety and makes the code more understandable, maintainable and less prone to errors.