Add getName method to filters and initializeValidityPeriod method to ScheduleManager
This commit adds a new `getName` method to every filter under the package `schedule.filter`, and an `initializeValidityPeriod` method to `ScheduleManager` in the `schedule.manager` package. The `getName` method allows for the retrieval of filter identifiers. The inclusion of `initializeValidityPeriod` eases the initialization of the schedule's validity period using properties from an external file. Further updates include: refinements to the `Schedule` and `ScheduleManagerJSON` classes, improving the program's overall usability and efficiency.