"Add Schedule initialization in Main class and enhance filter handling"
A new feature for initializing ScheduleManagerCollection and ScheduleManagerWeekly in Main class has been added so that the user can now choose either of them. This choice of schedule type is made interactive to improve user experience. Moreover, the enum HelpMenu has a new command type "INIT" to assist with the new schedule initialization feature.
Moreover, changes are introduced in the logic to handle 'filter' and 'exclusion' commands. This helps in managing these commands more effectively when the 'add' or 'remove' operations are performed. The filter creation logic has been streamlined for better efficiency. Filter creation is now handled by the Filter class enhancing code simplicity and maintenance.
The importSchedule method is rearranged to load classroom amenities before the schedule. A check is also removed that validates a scheduleConfigPath, as it seems unnecessary. These changes will enhance the performance of the importSchedule operation.